Delete custom NameID from the GUI oxShibboleth #39
