+++ oxService/src/main/java/org/xdi/config/oxtrust/AttributeResolverConfiguration.java
@@ -6,12 +6,14 @@
 
 package org.xdi.config.oxtrust;
 
+import java.util.List;
+
 import javax.enterprise.inject.Vetoed;
 
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 
 /**
- * Attribute resolver configuration
+ * Attribute resolver configurations
  * 
  * @author Yuriy Movchan Date: 09/04/2017
  */
@@ -19,41 +21,14 @@ import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AttributeResolverConfiguration implements Configuration {
 
-	private String attributeBase;
-	private String attributeName;
-	private String nameIdType;
-	private boolean enabled;
+	private List<NameIdConfig> nameIdConfigs;
 
-	public String getAttributeBase() {
-		return attributeBase;
-	}
+    public List<NameIdConfig> getNameIdConfigs() {
+        return nameIdConfigs;
+    }
 
-	public void setAttributeBase(String attributeBase) {
-		this.attributeBase = attributeBase;
-	}
-
-	public String getAttributeName() {
-		return attributeName;
-	}
-
-	public void setAttributeName(String attributeName) {
-		this.attributeName = attributeName;
-	}
-
-	public String getNameIdType() {
-		return nameIdType;
-	}
-
-	public void setNameIdType(String nameIdType) {
-		this.nameIdType = nameIdType;
-	}
-
-	public boolean isEnabled() {
-		return enabled;
-	}
-
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
+    public void setNameIdConfigs(List<NameIdConfig> nameIdConfigs) {
+        this.nameIdConfigs = nameIdConfigs;
+    }
 
 }
